Renamed vertical timeline mocks to match component name
