+++ src/__mocks__/verticalTimelineMocks.js
